feat: Swagger 경로 변경 및 Swagger Auth 함수 분리
